Implementing basic cluster alerting.
When having multiple grafana pointing a single database, alerts are
getting evaluated on each instances then notifications are dedup.
This patch allows any Grafana instance to set itself to own alerts
and be the only node to evaluate alerts. If that instances would
be stopped or die unexpectedly, the token would expire & another
instance would be able to take it.